requirements.txt
# Python dependencies will be added here
playwright
parsel
pymongo
pydantic
google-generativeai 
dotenv